Move MatrixMediaURL classes to api util
